requirements.txt
Authlib>=1.3.2
google-adk
google-genai
google-cloud-aiplatform
streamlit